Changes made in clinet.py
getDataPoint. In this method, you'll have to make modifications to compute the
right stock price. This means you have to change how `price` is computed by
using the formula: (bid_price+ask_price) / 2

getRatio. Right now, this method just returns 1 all the time. To correct this, you
must change the return value to the ratio of stock price_a to stock price_b

To review, we created a prices dictionary to store the stock prices. Think of
a dictionary as a key-value store wherein you can specify a key and be able
to retrieve a value. In our case, the key was the stock name and the value
was the price.
We then used this prices dictionary at the end to pass in the right values in
the getRatio function.